refactor(shadow dom): update shadow dom files to appropriate names
