+++ src/helix-ui/elements/HXAlertElement.less
@@ -1,4 +1,4 @@
-@import "./_hx-element";
+@import "./HXElement";
 @import "elements/hx-icon";
 
 button {
