+++ src/helix-ui/elements/HXCheckboxElement.less
@@ -1,4 +1,4 @@
-@import "./_hx-element";
+@import "./HXElement";
 @import "elements/hx-icon";
 
 #container {
