+++ src/helix-ui/elements/HXErrorElement.less
@@ -1,2 +1,2 @@
-@import './_hx-element';
+@import './HXElement';
 @import 'elements/hx-icon';
