+++ src/helix-ui/elements/HXPopoverElement.less
@@ -1,4 +1,4 @@
-@import "./_hx-element";
+@import "./HXElement";
 @import "./_positioning";
 
 #container {
