+++ src/helix-ui/elements/HXSearchElement.less
@@ -1,5 +1,5 @@
 @import (reference) "mixins";
-@import "./_hx-element.less";
+@import "./HXElement.less";
 @import "elements/hx-icon";
 
 :host {
