+++ src/helix-ui/elements/HXTooltipElement.less
@@ -1,4 +1,4 @@
-@import "./_hx-element";
+@import "./HXElement";
 @import "./_positioning";
 
 #container {
